# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SE-Textile Backend API
  description: Dynamically generated OpenAPI spec built from registered Flask routes.
  version: 2.0.0
servers:
- url: http://127.0.0.1:5001
  description: Local development server
security:
- bearerAuth: []
paths:
...
  /api/v1/catalog/load:
    post:
      tags:
      - catalog
      summary: Load fashion catalog data from CSV
      description: Reads the bundled catalog dataset, clears existing records, and reloads them into the database.
      responses:
        '200':
          description: Catalog loaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '404':
          description: Catalog file missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Failure while loading catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
  /api/v1/catalog/search:
    get:
      tags:
      - catalog
      summary: Search or filter catalog items
      description: Supports filtering by category, color, gender, season, usage, or keyword.
      responses:
        '200':
          description: Filtered catalog entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogSearchResponse'
        '500':
          description: Server error while filtering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: keyword
        in: query
        required: true
        schema:
          type: string
      - name: category
        in: query
        required: false
        schema:
          type: string
      - name: color
        in: query
        required: false
        schema:
          type: string
      - name: gender
        in: query
        required: false
        schema:
          type: string
      - name: season
        in: query
        required: false
        schema:
          type: string
      - name: usage
        in: query
        required: false
        schema:
          type: string
  /api/v1/catalog/view:
    get:
      tags:
      - catalog
      summary: View catalog items
      description: Returns a limited number of catalog entries.
      responses:
        '200':
          description: Catalog items returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogProduct'
        '500':
          description: Error fetching catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
...
  /api/v1/customer/discovery/nearby-shops:
    get:
      tags:
      - discovery_portal
      summary: Return shops located within a given radius from provided coordinates.
      description: Return shops located within a given radius from provided coordinates.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: lat
        in: query
        required: false
        schema:
          type: string
      - name: lon
        in: query
        required: false
        schema:
          type: string
      - name: radius
        in: query
        required: false
        schema:
          type: string
  /api/v1/customer/discovery/popular-shops:
    get:
      tags:
      - discovery_portal
      summary: Return popular shops with geolocation enrichment.
      description: Return popular shops with geolocation enrichment.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
  /api/v1/customer/discovery/search:
    get:
      tags:
      - discovery_portal
      summary: Search for fabrics or shops by keyword.
      description: Search for fabrics or shops by keyword.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
  /api/v1/customer/discovery/trending-fabrics:
    get:
      tags:
      - discovery_portal
      summary: Return top trending fabrics with AI-generated marketing captions.
      description: Return top trending fabrics with AI-generated marketing captions.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GenericSuccess:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
    GenericError:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        error:
          type: string
...
    CatalogProduct:
      type: object
      properties:
        product_id:
          type: string
        product_name:
          type: string
        category:
          type: string
        subcategory:
          type: string
        article_type:
          type: string
        color:
          type: string
        gender:
          type: string
        season:
          type: string
        usage:
          type: string
        image_url:
          type: string
        price:
          type: number
    CatalogSearchResponse:
      type: object
      properties:
        status:
          type: string
        total_results:
          type: integer
        products:
          type: array
          items:
            $ref: '#/components/schemas/CatalogProduct'
...